Move database setup from logger classes to app.rb
* Set up database connection in `app.rb' and introduce a constant
  that indicates the presence of a connection.
* Remove database initialization from logger, daily stats,
  and log cleaner classes; drop option for database URL from
  their initialize methods.
* Set up database only once for tests, by loading `app.rb'.

+++ test/helper.rb
@@ -2,18 +2,17 @@ require "minitest/autorun"
 
 DB_FILE = File.expand_path(File.dirname(__FILE__) + '/test.db')
 ENV["DATABASE_URL"] = "sqlite:///#{DB_FILE}"
+ENV["NO_LOGS"] = "true"
+
+require_relative "../app"
 
 
 def setup_database
-  DataMapper.auto_upgrade!
 end
 
 def teardown_database
-  Log.destroy
-  DailyStats.destroy
-
-  DataObjects::Pooling.pools.each {|pool| pool.dispose }  # close connection
-  File.delete(DB_FILE)  if File.exist?(DB_FILE)
+  Log.destroy  if DB
+  DailyStats.destroy  if DB
 end
